Add files via upload
A zombie simulator that simulates how many days it takes for a zombie to infect a village of 100 houses if the zombie does not know which houses are already infected and we make the assumption that one zombie takes one day to infect a house. We start on one infected house that is randomly placed, that house will then start infecting other houses.